Controllers update
Replaced old dtos for new dtos in the remaining controllers. Also done
some small fixes. PaginatedResponse attribute totalCost is of type int
now instead of long. Size of the passed collection is now passed in
PaginatedResponse in controllers instead of count method of repository
classes.

+++ tim7iss/src/main/java/com/tim7/iss/tim7iss/dto/PaginatedResponseDto.java
@@ -11,7 +11,7 @@ import java.util.Collection;
 @AllArgsConstructor
 public class PaginatedResponseDto<T> {
 
-    public Long totalCount;
+    public Integer totalCount;
     public Collection<T> results;
 
 }
